Add plain text Star notification mail
Prior to this commit, giving a Star would send an HTML-only mail.
This is unfortunate for people who are rocking text-ony mail clients,
like `mutt`, which is the best thing ever and you should try it.

This commit makes it so that Star notification mails are multipart;
they contain both 'text/plain' and 'text/html' parts, in the proper
order, so that both plain and rich email clients will display the
proper format.

Change-Id: I32a9635cca03784bcc55e2c055fb0575d8eb5cc3
Reviewed-on: https://gerrit.causes.com/22783
Reviewed-by: Elliot Block <elliot.block@causes.com>
Tested-by: Elliot Block <elliot.block@causes.com>