Gracefully handle invalid browser input for stars list
We were seeing the following error (modified for display purposes):

    Processing StarsController#create [POST]
    Parameters: {
      "star"=>{
        "to_ids"=>"[131]",
        "reason"=>"...",
        "star_type"=>"creative"
      }
    }

    NoMethodError (undefined method `reject' for "[131]":String):`)`

The `to_ids` parameter should be `["131"]`, not "[131]". This suggests
some browsers escape the input in such a way that prevents Rails from
interpreting the value correctly.

A quick but hacky solution is to check if the value is a string, and
attempt to convert it to a list by parsing the value as JSON.

Change-Id: Idf1af668cbe315b214729f77e86c07c9d8012568
Reviewed-on: https://gerrit.causes.com/32800
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>