Add /stars/recent endpoint
This endpoint can be used to export JSON or CSV data for stars given
during a certain timeframe. The result is ordered by number of stars and
grouped by the recipient(s). By default, stars for the last 30 days will
be returned. You can change that by passing in `Date#parse`:able strings
in `to` and `from` params, e.g.

/stars/recent?from=2014-02-27&to=2014-03-27

This will give you stars from February 27 to March 27.

Change-Id: I66d28a2d278c7ccf8e0c40cf914fbf476ae273d9
Reviewed-on: http://gerrit.causes.com/36613
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>